[refactor] #115 - 기존 MyFriend 컴포넌트 리팩토링
