+++ src/frontend/src/components/Modal/FriendModal/FriendComponent/FriendComponent.module.css
@@ -1,0 +1,71 @@
+.friendContainer {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  height: calc(61 - 20px);
+  padding: 10px 15px;
+  width: 100%;
+}
+
+.friendImg {
+  width: 35px;
+  height: 35px;
+  padding: 1px 0px;
+  gap: 10px;
+  border-radius: 300px;
+  background-color: var(--dark-gray);
+}
+
+.friendImg img {
+  width: 29px;
+  height: 29.5px;
+  margin-left: 3px;
+  margin-top: 3px;
+}
+
+.friendData {
+  display: flex;
+  flex-direction: column;
+  width: 197px;
+  height: calc(41 - 10px);
+  padding: 5px 10px;
+  gap: 5px;
+}
+
+.friendName {
+  width: 91px;
+  height: 14px;
+  font-weight: 600;
+  line-height: 14.32px;
+  text-align: left;
+  color: #000000;
+}
+
+.friendIntro {
+  width: 151px;
+  height: 12px;
+  font-weight: 400;
+  line-height: 11.93px;
+  text-align: left;
+  color: var(--dark-gray);
+}
+
+.friendFunction {
+  display: flex;
+  align-items: center;
+  width: 55px;
+  height: 25px;
+  padding: 8px 0px 8px 15px;
+  gap: 5px;
+}
+
+.friendFunction button {
+  background: none;
+  border: none;
+  color: "#000000";
+  width: 25px;
+  height: 24.66px;
+  padding: 3px 2px;
+  gap: 10px;
+  border-radius: 10px;
+}
